Add additional meta data to responses.
This MR will add two new attributes to response meta data
- `group`
- `statusCode`

This information is useful to parse this document and figure out the purpose of the responses. Here is an example of the output.

```
      "meta": {
        "type": "response",
        "title": null,
        "group": "Customer",
        "statusCode": "422"
      }
```

Just by looking at the metadata with another program I can tell the schema belongs to a response from Customer and is the response schema for status 422.

My use case is that the same endpoint can respond with different statuses and also have a different response format.